Fixing issues with longer timestamp strings

+++ rosbag/rgbd_benchmark_tools/associate.py
@@ -44,6 +44,7 @@ import argparse
 import sys
 import os
 import numpy
+from decimal import Decimal
 
 
 def read_file_list(filename):
@@ -65,7 +66,7 @@ def read_file_list(filename):
     data = file.read()
     lines = data.replace(","," ").replace("\t"," ").split("\n") 
     list = [[v.strip() for v in line.split(" ") if v.strip()!=""] for line in lines if len(line)>0 and line[0]!="#"]
-    list = [(float(l[0]),l[1:]) for l in list if len(l)>1]
+    list = [(Decimal(l[0]),l[1:]) for l in list if len(l)>1]
     return dict(list)
 
 def associate(first_list, second_list,offset,max_difference):
